Move UnpoweredFlashlight to shared (#27449)
