--- Content.Server/Light/Events/LightToggleEvent.cs
@@ -1,12 +1,0 @@
-namespace Content.Server.Light.Events
-{
-    public sealed class LightToggleEvent : EntityEventArgs
-    {
-        public bool IsOn;
-
-        public LightToggleEvent(bool isOn)
-        {
-            IsOn = isOn;
-        }
-    }
-}
