+++ Content.Shared/Actions/ActionEvents.cs
@@ -68,9 +68,10 @@ public sealed class GetItemActionsEvent : EntityEventArgs
         AddAction(ref actionId, prototypeId, Provider);
     }
 
-    public void AddAction(EntityUid actionId)
+    public void AddAction(EntityUid? actionId)
     {
-        Actions.Add(actionId);
+        if (actionId != null)
+            Actions.Add(actionId.Value);
     }
 }
