+++ Content.Shared/Light/LightToggleEvent.cs
@@ -1,0 +1,6 @@
+namespace Content.Shared.Light;
+
+public sealed class LightToggleEvent(bool isOn) : EntityEventArgs
+{
+    public bool IsOn = isOn;
+}
